refactor out the scheduler
