+++ lm_tournament_eval/api/offline_tournament.py
@@ -1,24 +1,10 @@
 # this is a collection of matches, models, and a schedule of "play"
 
 from dataclasses import dataclass, field
-#from offline_match_scheduler import OfflineMatchSchedulerConfig, OfflineMatchScheduler
+from .offline_match_scheduler import OfflineMatchSchedulerConfig, OfflineMatchScheduler
 import json
 import numpy as np
 
-@dataclass
-class OfflineMatchSchedulerConfig:
-    rounds : int
-    num_samples : int
-
-class OfflineMatchScheduler:
-    def __init__(self, config : OfflineMatchSchedulerConfig):
-        self.config = config
-
-    def schedule_tournament(self):
-        print("Creating Offline Schedule")
-        match_indices = np.random.choice(self.config.num_samples,
-        			         self.config.rounds,replace=False)
-        return match_indices
 
 
 @dataclass
